pyproject.toml
[tool.poetry]
name = "url-shortener-analytics-service"
version = "0.1.0"
description = ""
authors = ["Artem <Gakerchannel@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.13"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.0"
sqlalchemy = "^2.0.41"
asyncpg = "^0.30.0"
uvicorn = "^0.34.3"
black = "^25.1.0"
alembic = "1.14.0"
pyjwt = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
